requirements.txt
streamlit==1.38.0
openai==1.30.0
python-dotenv==1.0.1
pillow==10.4.0